[refactor] StoreSortBy 및 StoreFilterBy 클래스를 resolver 패키지에서 dto패키지로 이동
